+++ changelog.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## Version 1.0.1 (2021-11-15)
+## Version 1.0.0 (2021-11-15)
 
 ### Features
 
 - openapi `typescript-angular` generator custom templates for `@anglr/rest`
+- removal of `Using<HTTP_METHOD>` suffix for methods
+- replacement of *path* for methods with *path arguments*
+- working `ResponseTransform` adding